amended code in calcFeatureProbability() method by casting integer values to double
thus avoiding rounding caused by integer division that was previously
causing the test suite to produce errors

+++ TFS/src/twittercorpus/NaiveBayesClassifier.java
@@ -221,10 +221,10 @@ public class NaiveBayesClassifier implements Classifier {
             if(getTotalNumCategories() == 0){
                 throw new NoSuchElementException("No categories have been properly initialised!");
             } else {
-                return 1.0 / getTotalNumCategories();
+                return 1.0 / (double) getTotalNumCategories();
             }
         } else {
-            return fCountInCategory(feature,sentimentCategory) / featureTotalCount.get(feature);
+            return (double) fCountInCategory(feature,sentimentCategory) / (double) featureTotalCount.get(feature);
         }
     }
